Port numbers are now sent as decimal in both the port map message and the port change messages.  (bit mapped  mask value are  in hex as before)
Changed name of DEBUG_INFO_MSG_HEADER  constant to INFO_MSG_HEADER -  the protocol is unchanged (these messages still begin with !).  FYI,   low level debug messages (currently all disabled) will begin with and exclamation. Therefore a message that starts with !! will be displayed as in info message and is an indication that the developer (probably me) forget to turn of debugging in the library code.

Changed the way static flash memory strings are declared to fit the current Arduino style conventions.

Changed the way the class pointers are declared to improve cosmetics in the sketch (no more **pointer stuff in the sketch)